Auton Fixes
* Added AutoTurn() - The way the autodrive function was written, would
have caused the robot to try and hold an encoder count of zero while
turning, which would have caused a sort of pivot turn / arc turn rather
than a dead-spin. I don't think this is desirable, so I split the
function in two.
* Added absMax() - to limit drive motor speed, reduce typo potential,
increase code reuse.
* We never reset the gyro, so the auto turn feature would not have
worked as intended. That's ok, because if we avoid resetting the gyro we
get better overall accuracy, but the code needs to be updated to work
correctly.
* Apparently I deleted the limitBreak control last night... (fixed)