Post Esky
Ignore anything johnMode related, as it wasn't used at all. The problem with the final match is that we were shifted into high(and I didn't have shift to low code). I have no idea why the lift was dropping during the second leg of the auton as there was no code to tell it to do so. I think the main problem was with how  I screwed with the autoDrive function. I really shouldn't have had the allowed value to have been as high as 30 and just assumed it would coast to the location. Remember that one time we managed to pull off the switch auton but it was just millisecond too late? After that I increased the drive speed but that threw off the coasting values. Should I just decrease the allowed distance and tweak the kP value? Also, I don't like what I did with the autoTurn function, instead of tweaking the kP value because I didn't really know how to so I added a minimum speed function. One last thing, do you think there's any reason to have screen sharing? I have a discord & skype if you think it would be helpful. Thanks! :)